+++ StreRender/Resource/object_manager.cpp
@@ -93,9 +93,9 @@ std::wstring StringToWString(const std::string& str) {
 std::string WstringToString(std::wstring wstr)
 {
     std::string result;
-    int len = WideCharToMultiByte(CP_ACP, 0, wstr.c_str(), wstr.size(), NULL, 0, NULL, NULL);
+    int len = WideCharToMultiByte(CP_ACP, 0, wstr.c_str(), (int)wstr.size(), NULL, 0, NULL, NULL);
     char* buffer = new char[len + 1];
-    WideCharToMultiByte(CP_ACP, 0, wstr.c_str(), wstr.size(), buffer, len, NULL, NULL);
+    WideCharToMultiByte(CP_ACP, 0, wstr.c_str(), (int)wstr.size(), buffer, len, NULL, NULL);
     buffer[len] = '\0';
     result.append(buffer);
     delete[] buffer;
